requirements.txt
telethon==1.36.0
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
aiosqlite==0.20.0
httpx==0.27.2
python-multipart==0.0.9
deep-translator==1.11.4
langdetect==1.0.9
